Add DatabaseCleaner::strategy getter method
Issue #511 proposes the addition of a ::strategy getter
on the DatabaseCleaner module to aid new users with debugging.
This commit introduces that method.

Like the ::strategy= setter method, ::strategy is actually defined
as an instance method on DatabaseCleaner::Cleaners,
and is made available to the top-level module by delegation:

* DatabaseCleaner::strategy delegates to
* DatabaseCleaner::Configuration#strategy, which delegates to
* DatabaseCleaner::Cleaners#strategy.

Depending on the complexity of the user's configuration,
this method may return three different kinds of values:

1. nil (if no strategies have been set on any cleaners);
2. a single strategy (if all cleaners share the same strategy); or
3. a hash of cleaner IDs and the strategies that go with them
   (if multiple strategies have been set on multiple cleaners).

The first of these three cases has not explicitly been tested.